move tasks and personality to utils folder
