+++ App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Task from "./Task";
+import Task from "./src/components/utils/Task";
 import AppNavigator from "./src/navigation/AppNavigator";
 import { ThemeProvider} from "react-native-rapi-ui";
 export default function App() {
