+++ src/components/utils/Task.js
@@ -1,5 +1,5 @@
 import React from "react"
-import AppNavigator from "./src/navigation/AppNavigator";
+import AppNavigator from "../../navigation/AppNavigator";
 import { ThemeProvider} from "react-native-rapi-ui";
 
 // the two arrays that contain all tasks
